# Conflicts:
#	cloudprinter/src/main/java/com/sunmi/cloudprinter/ui/activity/PrinterManageActivity.java
#	ipc/src/main/res/values-zh-rCN/strings.xml
#	ipc/src/main/res/values/strings.xml

+++ app/src/main/java/com/sunmi/assistant/mine/shop/ShopDetailActivity.java
@@ -228,7 +228,7 @@ public class ShopDetailActivity extends BaseActivity {
     public void onAreaResult(int resultCode, Intent data) {
         if (resultCode == Activity.RESULT_OK) {
             isUpdateShopInfo = true;
-            info.setBusinessArea(Float.parseFloat(data.getStringExtra(INTENT_EXTRA_AREA)));
+            info.setBusinessArea(Double.parseDouble(data.getStringExtra(INTENT_EXTRA_AREA)));
             silShopArea.setEndContent(floatTrans(info.getBusinessArea()) + "㎡");
         }
     }
